m3. cr-fixed: renamed some files, add file game.logic.js, ..and other fixed
