+++ js/main.js
@@ -1,4 +1,4 @@
-import {showScreen} from './util';
+import {showScreen} from './utils';
 import intro from './screen-00-intro';
 
 showScreen(intro);
